src/PostService/requirements.txt
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
grpcio==1.40.0
grpcio-tools==1.40.0
python-dotenv==0.19.0
